Fix CLI semantic supersession queries
