--- .github/workflows/fix-cli-supersession.yml
@@ -1,45 +1,0 @@
-name: Fix and verify CLI supersession
-
-on:
-  push:
-    branches:
-      - fix-cli-supersession-query
-
-permissions:
-  contents: write
-
-jobs:
-  fix-and-test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-      - uses: dtolnay/rust-toolchain@stable
-        with:
-          components: rustfmt
-      - name: Apply source changes
-        run: |
-          python3 scripts/apply_cli_supersession_fix.py
-          cargo fmt
-      - name: Run regression test
-        run: cargo test --test cli_supersession -- --nocapture
-      - name: Run library tests
-        run: cargo test --lib
-      - name: Run real adversarial CLI check
-        run: |
-          rm -rf /tmp/lint-ai-supersession-proof .lint-ai-cache
-          mkdir -p /tmp/lint-ai-supersession-proof
-          printf '# Gateway Retry Policy\n\nGateway timeout retry attempts: 5.\n' > /tmp/lint-ai-supersession-proof/decision-a.md
-          printf '%s\n' '---' 'supersedes: decision-a.md' '---' '# Gateway Retry Policy' '' 'Gateway timeout retry attempts: 2.' > /tmp/lint-ai-supersession-proof/decision-b.md
-          cargo run --quiet --bin lint-ai -- --query 'How many retry attempts should we use for gateway timeouts?' /tmp/lint-ai-supersession-proof > /tmp/query.json
-          cargo run --quiet --bin lint-ai -- --llm-context 'How many retry attempts should we use for gateway timeouts?' --result-count 5 /tmp/lint-ai-supersession-proof > /tmp/context.json
-          python3 -c "import json; q=json.load(open('/tmp/query.json')); c=json.load(open('/tmp/context.json')); assert [r['doc_id'] for r in q['results']]==['decision-b.md'],q; assert q['results'][0].get('semantic_status')=='current',q; assert q.get('aggregation') is None,q; assert all(x['doc_id']!='decision-a.md' for x in c['top_chunks']),c; assert any(x['doc_id']=='decision-b.md' and 'retry attempts: 2' in x['text'] for x in c['top_chunks']),c; print(json.dumps({'results':q['results'],'aggregation':q.get('aggregation'),'top_chunks':c['top_chunks']},indent=2))"
-      - name: Commit verified fix
-        run: |
-          rm .github/workflows/fix-cli-supersession.yml scripts/apply_cli_supersession_fix.py
-          git config user.name 'RooAGI Developer'
-          git config user.email 'rooagidev@gmail.com'
-          git add -A
-          git commit -m 'Fix CLI semantic supersession queries'
-          git push origin HEAD:fix-cli-supersession-query
